Register Go stores automatically, with optional enablement
