+++ mlflow_go/store/model_registry.py
@@ -8,8 +8,9 @@ from mlflow.protos.model_registry_pb2 import (
     GetLatestVersions,
 )
 
+from mlflow_go import is_go_enabled
 from mlflow_go.lib import get_lib
-from mlflow_go.store import ServiceProxy
+from mlflow_go.store._service_proxy import _ServiceProxy
 
 _logger = logging.getLogger(__name__)
 
@@ -23,7 +24,7 @@ class _ModelRegistryStore:
                 "log_level": logging.getLevelName(_logger.getEffectiveLevel()),
             }
         ).encode("utf-8")
-        self.service = ServiceProxy(get_lib().CreateModelRegistryService(config, len(config)))
+        self.service = _ServiceProxy(get_lib().CreateModelRegistryService(config, len(config)))
         super().__init__(store_uri)
 
     def __del__(self):
@@ -43,3 +44,12 @@ class _ModelRegistryStore:
 
 def ModelRegistryStore(cls):
     return type(cls.__name__, (_ModelRegistryStore, cls), {})
+
+
+def _get_sqlalchemy_store(store_uri):
+    from mlflow.store.model_registry.sqlalchemy_store import SqlAlchemyStore
+
+    if is_go_enabled():
+        SqlAlchemyStore = ModelRegistryStore(SqlAlchemyStore)
+
+    return SqlAlchemyStore(store_uri)
